+++ src/woxml.rs
@@ -1,148 +1,22 @@
 // Copyright © 2025 Stephan Kunz
 // Copyright © of xml_writer::XmlWriter Piotr Zolnierek
 
-extern crate alloc;
-
 use alloc::{boxed::Box, string::ToString, vec::Vec};
 use core::fmt::{Debug, Formatter};
 
-use thiserror::Error;
+use crate::{
+	error::{Error, Result},
+	write::Write,
+};
 
-pub type Result<T> = core::result::Result<T, Error>;
-
-/// Error type
-#[derive(Error, Debug)]
-pub enum Error {
-	/// Pass through `core::str::Utf8Error`
-	#[error("{0}")]
-	Utf8(#[from] core::str::Utf8Error),
-	/// Pass through `core::fmt::Error`
-	#[error("{0}")]
-	Core(#[from] core::fmt::Error),
-	/// failed to write whole buffer
-	#[error("failed to write whole buffer")]
-	WriteAllEof,
-}
-
-/// The trait for objects which are byte-oriented sinks.
-pub trait Write {
-	/// Flushes this output stream, ensuring that all intermediately buffered contents reach their destination.
-	/// # Errors
-	/// It is considered an error if not all bytes could be written due to I/O errors or EOF being reached.
-	fn flush(&mut self) -> Result<()>;
-
-	/// Writes a buffer into this writer, returning how many bytes were written.
-	///
-	/// # Errors
-	/// Each call to write may generate an I/O error indicating that the operation could not be completed. If an error is returned then no bytes in the buffer were written to this writer.
-	/// It is not considered an error if the entire buffer could not be written to this writer.
-	fn write(&mut self, buf: &[u8]) -> Result<usize>;
-
-	/// Attempts to write an entire buffer into this writer.
-	///
-	/// This method will continuously call write until there is no more data to be written.
-	/// This method will not return until the entire buffer has been successfully written or an error occurs.
-	/// # Errors
-	/// This function will return the first error that write returns.
-	fn write_all(&mut self, mut buf: &[u8]) -> Result<()> {
-		while !buf.is_empty() {
-			match self.write(buf) {
-				Ok(0) => {
-					return Err(Error::WriteAllEof);
-				}
-				Ok(n) => buf = &buf[n..],
-				//Err(ref e) => e.is_interrupted() => {}
-				Err(e) => return Err(e),
-			}
-		}
-		Ok(())
-	}
-}
-
-// #[cfg(not(feature = "std"))]
-impl Write for Vec<u8> {
-	#[inline]
-	fn flush(&mut self) -> Result<()> {
-		Ok(())
-	}
-
-	#[inline]
-	fn write(&mut self, buf: &[u8]) -> Result<usize> {
-		self.extend_from_slice(buf);
-		Ok(buf.len())
-	}
-
-	#[inline]
-	fn write_all(&mut self, data: &[u8]) -> Result<()> {
-		if self.write(data)? < data.len() {
-			Err(Error::WriteAllEof)
-		} else {
-			Ok(())
-		}
-	}
-}
-
-// #[cfg(not(feature = "std"))]
-impl Write for bytes::BytesMut {
-	#[inline]
-	fn flush(&mut self) -> Result<()> {
-		Ok(())
-	}
-
-	#[inline]
-	fn write(&mut self, buf: &[u8]) -> Result<usize> {
-		self.extend_from_slice(buf);
-		Ok(buf.len())
-	}
-
-	#[inline]
-	fn write_all(&mut self, data: &[u8]) -> Result<()> {
-		if self.write(data)? < data.len() {
-			Err(Error::WriteAllEof)
-		} else {
-			Ok(())
-		}
-	}
-}
-
-// /// Implement the trait for all types that implement [`core::fmt::Write`]
-// #[cfg(not(feature = "std"))]
-// impl<T> Write for T
-// where
-//     T: core::fmt::Write,
-// {
-//     #[inline]
-//     fn flush(&mut self) -> Result<()> {
-//         Ok(())
-//     }
-
-//     #[inline]
-//     fn write(&mut self, buf: &[u8]) -> Result<usize> {
-//         let str = str::from_utf8(buf)?;
-//         let len = str.len();
-//         T::write_str(self, str)?;
-//         Ok(len)
-//     }
-// }
-
-// /// Implement the trait for all types that implement [`std::io::Write`]
-// #[cfg(feature = "std")]
-// impl<T> Write for T
-// where
-//     T: std::io::Write,
-// {
-//     #[inline]
-//     fn flush(&mut self) -> Result<()> {
-//         T::flush(self)?;
-//         Ok(())
-//     }
-
-//     #[inline]
-//     fn write(&mut self, buf: &[u8]) -> Result<usize> {
-//         let len = T::write(self, buf)?;
-//         Ok(len)
-//     }
-// }
+/// Multiple used literal definitions
+const CLOSE: &str = ">";
+const CLOSE_CLOSE: &str = "/>";
+const OPEN: &str = "<";
+const SELF_CLOSE_OPEN: &str = "</";
+const SPACE: &str = " ";
+const EQUAL_QUOTE: &str = "=\"";
+const QUOTE: &str = "\"";
 
 /// The `XmlWriter` himself.
 pub struct XmlWriter<'a, Buf: Write> {
@@ -167,7 +41,11 @@ pub struct XmlWriter<'a, Buf: Write> {
 
 impl<Buf: Write> Debug for XmlWriter<'_, Buf> {
 	fn fmt(&self, f: &mut Formatter) -> core::fmt::Result {
-		Ok(write!(f, "XmlWriter {{ stack: {:?}, opened: {} }}", self.stack, self.opened)?)
+		Ok(write!(
+			f,
+			"XmlWriter {{ stack: {:?}, namespaces: {:?}, opened: {} }}",
+			self.stack, self.ns_stack, self.opened
+		)?)
 	}
 }
 
@@ -260,14 +138,11 @@ impl<'a, W: Write> XmlWriter<'a, W> {
 	/// Writes namespace declarations (xmlns:xx) into the currently open element.
 	/// # Errors
 	/// - if writing to buffer fails
-	/// # Panics
 	/// - when opening a namespace without having an element
 	pub fn ns_decl(&mut self, ns_map: &Vec<(Option<&'a str>, &'a str)>) -> Result<()> {
-		assert!(
-			self.opened,
-			"Attempted to write namespace decl to elem, when no elem was opened, stack {:?}",
-			self.stack
-		);
+		if !self.opened {
+			return Err(Error::OpenNamespaceWithoutElement);
+		}
 
 		for item in ns_map {
 			let name = item
@@ -284,11 +159,11 @@ impl<'a, W: Write> XmlWriter<'a, W> {
 	pub fn elem(&mut self, name: &str) -> Result<()> {
 		self.close_elem(false)?;
 		self.indent()?;
-		self.write("<")?;
+		self.write(OPEN)?;
 		let ns = self.namespace;
 		self.ns_prefix(ns)?;
 		self.write(name)?;
-		self.write("/>")
+		self.write(CLOSE_CLOSE)
 	}
 
 	/// Write an element with inlined text content (escaped)
@@ -297,17 +172,17 @@ impl<'a, W: Write> XmlWriter<'a, W> {
 	pub fn elem_text(&mut self, name: &str, text: &str) -> Result<()> {
 		self.close_elem(false)?;
 		self.indent()?;
-		self.write("<")?;
+		self.write(OPEN)?;
 		let ns = self.namespace;
 		self.ns_prefix(ns)?;
 		self.write(name)?;
-		self.write(">")?;
+		self.write(CLOSE)?;
 
 		self.escape(text, false)?;
 
-		self.write("</")?;
+		self.write(SELF_CLOSE_OPEN)?;
 		self.write(name)?;
-		self.write(">")
+		self.write(CLOSE)
 	}
 
 	/// Begin an elem, make sure name contains only allowed chars
@@ -323,7 +198,7 @@ impl<'a, W: Write> XmlWriter<'a, W> {
 		self.indent()?;
 		self.stack.push((name, false));
 		self.ns_stack.push(self.namespace);
-		self.write("<")?;
+		self.write(OPEN)?;
 		self.opened = true;
 		// stderr().write_fmt(format_args!("\nbegin {}", name));
 		let ns = self.namespace;
@@ -337,9 +212,9 @@ impl<'a, W: Write> XmlWriter<'a, W> {
 	fn close_elem(&mut self, has_children: bool) -> Result<()> {
 		if self.opened {
 			if self.pretty && !has_children {
-				self.write("/>")?;
+				self.write(CLOSE_CLOSE)?;
 			} else {
-				self.write(">")?;
+				self.write(CLOSE)?;
 			}
 			self.opened = false;
 		}
@@ -349,16 +224,13 @@ impl<'a, W: Write> XmlWriter<'a, W> {
 	/// End and elem
 	/// # Errors
 	/// - if writing to buffer fails
-	/// # Panics
 	/// - when trying to close a namespace without having one opened
+	/// - when trying to close an element without having one opened
 	pub fn end_elem(&mut self) -> Result<()> {
 		self.close_elem(false)?;
-		let ns = self.ns_stack.pop().unwrap_or_else(|| {
-			panic!(
-				"Attempted to close namespaced element without corresponding open namespace, stack {:?}",
-				self.ns_stack
-			)
-		});
+		let Some(ns) = self.ns_stack.pop() else {
+			return Err(Error::CloseNamespace);
+		};
 		match self.stack.pop() {
 			Some((name, children)) => {
 				if self.pretty {
@@ -371,13 +243,13 @@ impl<'a, W: Write> XmlWriter<'a, W> {
 					}
 					self.newline = true;
 				}
-				self.write("</")?;
+				self.write(SELF_CLOSE_OPEN)?;
 				self.ns_prefix(ns)?;
 				self.write(name)?;
-				self.write(">")?;
+				self.write(CLOSE)?;
 				Ok(())
 			}
-			None => panic!("Attempted to close an elem, when none was open, stack {:?}", self.stack),
+			None => Err(Error::CloseElement),
 		}
 	}
 
@@ -392,48 +264,42 @@ impl<'a, W: Write> XmlWriter<'a, W> {
 			self.stack.push(previous);
 		}
 		self.indent()?;
-		self.write("<")?;
+		self.write(OPEN)?;
 		let ns = self.namespace;
 		self.ns_prefix(ns)?;
 		self.write(name)?;
-		self.write("/>")
+		self.write(CLOSE_CLOSE)
 	}
 
 	/// Write an attr, make sure name and value contain only allowed chars.
 	/// For an escaping version use `attr_esc`
 	/// # Errors
 	/// - if writing to buffer fails
-	/// # Panics
 	/// - when writing attributes without having an element
 	pub fn attr(&mut self, name: &str, value: &str) -> Result<()> {
-		assert!(
-			self.opened,
-			"Attempted to write attr to elem, when no elem was opened, stack {:?}",
-			self.stack
-		);
-		self.write(" ")?;
+		if !self.opened {
+			return Err(Error::WriteWithoutElement);
+		}
+		self.write(SPACE)?;
 		self.write(name)?;
-		self.write("=\"")?;
+		self.write(EQUAL_QUOTE)?;
 		self.write(value)?;
-		self.write("\"")
+		self.write(QUOTE)
 	}
 
 	/// Write an attr, make sure name contains only allowed chars.
 	/// # Errors
 	/// - if writing to buffer fails
-	/// # Panics
 	/// - when writing attributes without having an element
 	pub fn attr_esc(&mut self, name: &str, value: &str) -> Result<()> {
-		assert!(
-			self.opened,
-			"Attempted to write attr to elem, when no elem was opened, stack {:?}",
-			self.stack
-		);
-		self.write(" ")?;
+		if !self.opened {
+			return Err(Error::WriteWithoutElement);
+		}
+		self.write(SPACE)?;
 		self.escape(name, true)?;
-		self.write("=\"")?;
+		self.write(EQUAL_QUOTE)?;
 		self.escape(value, false)?;
-		self.write("\"")
+		self.write(QUOTE)
 	}
 
 	/// Escape identifiers or text.