+++ CHANGELOG.md
@@ -16,6 +16,19 @@ especially the [Rust flavour](https://doc.rust-lang.org/cargo/reference/semver.h
 
 ### Removed
 
+## [0.4.0] - 2025-??-??
+
+### Added
+
+### Changed
+- Error messages are non_exhaustive
+- no longer panics but errors when closing more elements/namespaces than opened
+- no longer panics but errors when writing without an open element
+
+### Fixed
+
+### Removed
+
 ## [0.3.0] - 2025-08-05
 
 ### Added